Update file structure to Testing 4
