+++ types/foundry/client-esm/canvas/module.d.ts
@@ -1,3 +1,4 @@
 export * as edges from "./edges/module.ts";
+export * as regions from "./regions/module.ts";
 export * from "./scene-manager.ts";
 export * as sources from "./sources/module.ts";
