+++ types/foundry/client-esm/canvas/regions/geometry.d.ts
@@ -1,0 +1,30 @@
+import type { Region } from "../../../client/pixi/placeables/region.d.ts";
+
+/**
+ * The geometry of a {@link Region}.
+ * - Vertex Attribute: `aVertexPosition` (`vec2`)
+ * - Draw Mode: `PIXI.DRAW_MODES.TRIANGLES`
+ */
+export class RegionGeometry extends PIXI.Geometry {
+    /**
+     * Create a RegionGeometry.
+     * @param    region    The Region to create the RegionGeometry from.
+     * @internal
+     */
+    constructor(region: Region);
+
+    /** The Region this geometry belongs to. */
+    get region(): Region;
+
+    /**
+     * Update the buffers.
+     * @internal
+     */
+    _clearBuffers(): void;
+
+    /**
+     * Update the buffers.
+     * @internal
+     */
+    _updateBuffers(): void;
+}
