+++ types/foundry/client-esm/canvas/regions/module.d.ts
@@ -1,0 +1,4 @@
+export * from "./geometry.ts";
+export * from "./mesh.ts";
+export * from "./polygon-tree.ts";
+export * from "./shape.ts";
