+++ types/foundry/client/pixi/webgl/shaders/index.d.ts
@@ -1,1 +1,2 @@
+import "./region.d.ts";
 import "./sampler.d.ts";
