+++ types/foundry/client/pixi/webgl/shaders/region.d.ts
@@ -1,0 +1,6 @@
+export {};
+
+/** The shader used by {@link RegionMesh}. */
+declare global {
+    class RegionShader extends AbstractBaseShader {}
+}
